refactor(webapp): align tagged items with container/content convention
Replace TaggedItemsLayout and the three per-type containers with a single
TaggedItemsContainer that owns the chrome (results header, TagFilters) and
the Suspense boundary, plus one ContainerContent per item type that fetches
and renders its own empty state and infinite scroll, matching the rest of
the app. Filters now stay mounted while the list suspends, the skeleton is
no longer borrowed from the search feature, and empty-state copy is
per item type.
