+++ src/webapp/app/(dashboard)/tags/[tag]/collections/page.tsx
@@ -1,6 +1,4 @@
-import TaggedCollectionsContainer from '@/features/tags/containers/taggedCollectionsContainer/TaggedCollectionsContainer';
-import CardSearchResultsContainerSkeleton from '@/features/search/containers/cardSearchResultsContainer/Skeleton.CardSearchresultsContainerSkeleton';
-import { Suspense } from 'react';
+import TaggedItemsContainer from '@/features/tags/containers/taggedItemsContainer/TaggedItemsContainer';
 
 interface Props {
   params: Promise<{ tag: string }>;
@@ -13,11 +11,10 @@ export default async function Page(props: Props) {
   const decoded = decodeURIComponent(tag).toLowerCase();
 
   return (
-    <Suspense
-      fallback={<CardSearchResultsContainerSkeleton />}
-      key={`${decoded}|${handle ?? ''}`}
-    >
-      <TaggedCollectionsContainer tag={decoded} handleOrDid={handle} />
-    </Suspense>
+    <TaggedItemsContainer
+      tag={decoded}
+      itemType="collection"
+      handleOrDid={handle}
+    />
   );
 }
